backend/requirements.txt
async-timeout==4.0.2
bcrypt==3.2.2
black==22.3.0
boto3==1.24.84
botocore==1.27.84
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
distlib==0.3.4
filelock==3.7.1
Flask==2.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
flask-mail==0.9.1
identify==2.5.1
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.19.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
redis==4.2.2
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.1
wrapt==1.14.1
zipp==3.8.0
